+++ medaka/stitch.py
@@ -14,16 +14,21 @@ import medaka.datastore
 import medaka.labels
 
 
-def write_fasta(filename, contigs):
-    """Write a fasta file from tuples of (name, sequence).
+def write_fastx_segment(fh, contig, qualities=True):
+    """Write a fastx file from a tuple of (name, sequence, qualities).
 
-    :param filename: output filename.
-    :param contigs: tuples of the form (sequence name, base sequence).
+    :param filename: output filehandle.
+    :param contigs: tuple of the form (sequence name, base sequence,
+        base qualities).
 
     """
-    with open(filename, 'w') as fasta:
-        for name, seq in contigs:
-            fasta.write('>{}\n{}\n'.format(name, seq))
+    if qualities:
+        fastx_prefix = '@'
+    else:
+        fastx_prefix = '>'
+    fh.write('{}{}\n{}\n'.format(fastx_prefix, contig[0], ''.join(contig[1])))
+    if qualities:
+        fh.write('+\n{}\n'.format(''.join(contig[2])))
 
 
 def stitch_from_probs(h5_fp, region):
@@ -36,7 +41,7 @@ def stitch_from_probs(h5_fp, region):
     :param h5_fp: iterable of HDF5 filepaths.
     :param region: `medaka.common.Region` instance
 
-    :returns: list of (region string, sequence).
+    :returns: list of (region string, sequence, qualities).
     """
     logger = medaka.common.get_named_logger('Stitch')
     logger.debug("Stitching region: {}".format(str(region)))
@@ -52,23 +57,29 @@ def stitch_from_probs(h5_fp, region):
     cur_ref_name = None
     heuristic_use = 0
     seq_parts = list()
+    qualities = list()
     start = None
     contigs = []
     for s, is_last_in_contig, heuristic in data_gen:
         cur_ref_name = s.ref_name if cur_ref_name is None else cur_ref_name
         start = s.positions[0]['major'] if start is None else start
-        seq_parts.append(label_scheme.decode_consensus(s))
+        seq, qual = label_scheme.decode_consensus(s, with_qualities=True)
+        seq_parts.append(seq)
+        qualities.append(qual)
         if is_last_in_contig:
             contigs.append((
                 (s.ref_name, start, s.positions[-1]['major']),
-                seq_parts))
+                seq_parts,
+                qualities))
             seq_parts = list()
+            qualities = list()
             start = None
         heuristic_use += heuristic
     if len(seq_parts) > 0:
         contigs.append((
             (s.ref_name, start, s.positions[-1]['major']),
-            seq_parts))
+            seq_parts,
+            qualities))
 
     logger.debug("Used heuristic {} times for {}.".format(
         heuristic_use, region))
@@ -78,7 +89,8 @@ def stitch_from_probs(h5_fp, region):
 def fill_gaps(contigs, draft):
     """Fill gaps between polished contigs with draft sequence.
 
-    :param contigs: iterable of ((ref_name, start, stop), sequence parts)
+    :param contigs: iterable of ((ref_name, start, stop), sequence parts,
+        qualities)
     :param draft: `pysam.FastaFile` or filepath of draft sequence.
 
     :returns: iterable of (name, info, seq)
@@ -88,12 +100,13 @@ def fill_gaps(contigs, draft):
 
     contig_trees = collections.defaultdict(intervaltree.IntervalTree)
     ordered_contigs = collections.OrderedDict()
-    for info, sequence_parts in contigs:
+    for info, sequence_parts, qualities in contigs:
         ref_name, start, stop = info
         # add one to end of interval, as intervaltree intervals and bed file
         # intervals are end-exclusive (i.e. they don't contain the last
         # coordinate), whilst the last position in a sample is included.
-        contig_trees[ref_name].addi(start, stop + 1, data=sequence_parts)
+        contig_trees[ref_name].addi(start, stop + 1, data=(
+            sequence_parts, qualities))
         ordered_contigs[ref_name] = None
 
     contig_lengths = dict(zip(draft.references, draft.lengths))
@@ -107,17 +120,22 @@ def fill_gaps(contigs, draft):
     for ref_name in ordered_contigs:
         contig_trees[ref_name].update(gap_trees[ref_name])
         draft_seq = draft.fetch(ref_name)
-        pieces = []
+        seq_pieces = []
+        qual_pieces = []
         for i in sorted(
                 contig_trees[ref_name], key=operator.attrgetter('begin')):
             if i.data is None:  # this is a gap
                 seq = [draft_seq[i.begin: i.end]]
+                qual = ['!' * (i.end - i.begin)]  # set Q=0 for padding bases
             else:
-                seq = i.data
-            pieces.extend(seq)
+                seq = i.data[0]
+                qual = i.data[1]
+            seq_pieces.extend(seq)
+            qual_pieces.extend(qual)
         # return as input
         stitched_contigs.append((
-            (ref_name, 0, contig_lengths[ref_name]), pieces))
+            (ref_name, 0, contig_lengths[ref_name]), seq_pieces,
+            qual_pieces))
     return stitched_contigs, gap_trees
 
 
@@ -132,18 +150,21 @@ def collapse_neighbours(contigs):
         return
     ref_name, start, stop = contig[0]
     buffer = contig[1]
+    qual = contig[2]
     for contig in contigs:
         c_rn, c_start, c_stop, = contig[0]
         if c_rn == ref_name and c_start == stop + 1:
             stop = c_stop
             buffer.extend(contig[1])
+            qual.extend(contig[2])
         else:
             # clear buffer, start anew
             yield (
-                (ref_name, start, stop), buffer)
+                (ref_name, start, stop), buffer, qual)
             ref_name, start, stop = contig[0]
             buffer = contig[1]
-    yield ((ref_name, start, stop), buffer)
+            qual = contig[2]
+    yield ((ref_name, start, stop), buffer, qual)
 
 
 def stitch(args):
@@ -182,7 +203,7 @@ def stitch(args):
         for r in regions_to_process))
 
     gap_trees = {}
-    with open(args.output, 'w') as fasta:
+    with open(args.output, 'w') as fastx:
         Executor = concurrent.futures.ProcessPoolExecutor
         with Executor(max_workers=args.threads) as executor:
             worker = functools.partial(stitch_from_probs, args.inputs)
@@ -194,24 +215,25 @@ def stitch(args):
                 # TODO: ideally fill_gaps would be a generator
                 contigs, gt = fill_gaps(contigs, args.draft)
                 gap_trees.update(gt)
-                for (ref_name, start, stop), seq_parts in contigs:
-                    fasta.write(">{}\n".format(ref_name))
-                    for s in seq_parts:
-                        fasta.write(s)
-                    fasta.write("\n")
+                for (ref_name, start, stop), seq_parts, qualities in contigs:
+                    write_fastx_segment(
+                        fastx,
+                        (ref_name, seq_parts, qualities),
+                        qualities=args.qualities)
             else:
                 ref_name = None
                 counter = 0
-                for (rname, start, stop), seq_parts in contigs:
+                for (rname, start, stop), seq_parts, qualities in contigs:
                     if ref_name == rname:
                         counter += 1
                     else:
                         counter = 0
-                    fasta.write(">{}_{} {}-{}\n".format(
-                        rname, counter, start, stop + 1))
-                    for s in seq_parts:
-                        fasta.write(s)
-                    fasta.write("\n")
+                    write_fastx_segment(
+                        fastx,
+                        ("{}_{} {}-{}".format(
+                            rname, counter, start, stop + 1),
+                         seq_parts, qualities),
+                        qualities=args.qualities)
                     ref_name = rname
         # the data index doesn't contain entries for contigs that
         # were entirely skipped (and we explicitely removed these
@@ -225,7 +247,10 @@ def stitch(args):
                 logger.info(
                     "Copying contig '{}' verbatim from input.".format(reg))
                 seq = draft.fetch(reg)
-                fasta.write(">{}\n{}\n".format(reg, seq))
+                write_fastx_segment(
+                    fastx,
+                    (reg, seq, '!' * len(seq)),
+                    qualities=args.qualities)
                 tree = intervaltree.IntervalTree()
                 tree.addi(0, draft_lengths[reg])
                 gap_trees[reg] = tree